Aggregate pong namespaces
